Move `SharedObject` to `types::shared_object` module
